Update to 20160215 release.
byuu says:

Got it. Wow, that didn't hurt nearly as much as I thought it was going
to.

Dropped from 127.5fps to 123.5fps to use Natural/Integer for
(u)int(8,16,32,64).

That's totally worth the cost.

+++ higan/emulator/emulator.cpp
@@ -17,9 +17,9 @@ auto Interface::videoColor(uint16 r, uint16 g, uint16 b) -> uint32 {
 
   if(gamma != 1.0) {
     double reciprocal = 1.0 / 32767.0;
-    r = r > 32767 ? r : 32767 * pow(r * reciprocal, gamma);
-    g = g > 32767 ? g : 32767 * pow(g * reciprocal, gamma);
-    b = b > 32767 ? b : 32767 * pow(b * reciprocal, gamma);
+    r = r > 32767 ? r : uint16(32767 * pow(r * reciprocal, gamma));
+    g = g > 32767 ? g : uint16(32767 * pow(g * reciprocal, gamma));
+    b = b > 32767 ? b : uint16(32767 * pow(b * reciprocal, gamma));
   }
 
   if(luminance != 1.0) {
