+++ higan/processor/v30mz/v30mz.cpp
@@ -250,10 +250,10 @@ auto V30MZ::instruction() -> void {
   case 0xcd: return opIntImm();
   case 0xce: return opInto();
   case 0xcf: return opReturnInt();
-  case 0xd0: return opGroup2MemImm(Byte, 1);
-  case 0xd1: return opGroup2MemImm(Word, 1);
-  case 0xd2: return opGroup2MemImm(Byte, r.cl);
-  case 0xd3: return opGroup2MemImm(Word, r.cl);
+  case 0xd0: return opGroup2MemImm(Byte, (uint8)1);
+  case 0xd1: return opGroup2MemImm(Word, (uint8)1);
+  case 0xd2: return opGroup2MemImm(Byte, (uint8)r.cl);
+  case 0xd3: return opGroup2MemImm(Word, (uint8)r.cl);
   case 0xd4: return opAdjustAfterMultiply();
   case 0xd5: return opAdjustAfterDivide();
   case 0xd6: return;
